Updated file paths for modularization changes
